Reformatted command arguments matrix
Modelled off of Table 2.3-2 in the PF working doc: https://github.com/oasis-tcs/openc2-ap-pf/blob/working/oc2pf.md#23-openc2-commands

In regards to "response_requested": Are there cases where this _isn't_ a valid argument? From how I read the LS, this doesn't seem to be the case. If so, I'd like to move that argument out of the matrix and mention it somewhere else. This would make the matrix look much cleaner.