Changes to denying/allowing network connections
Previously, the AP referred allow/deny IPv4/6 Commands to the SLPF AP. This is incorrect. EDR systems do not block traffic the way firewalls do. Some change the etc\hosts file or routing table to point back to localhost, others intervene with the connecting process by crashing it or otherwise halting the process. 

Point is that it looks at system events, not packets, and that concepts like ingress/egress, stateful/stateless and similar from the Packet Filtering AP do not apply here. This is just a simple blacklist for network locations.

These commands have therefore been changed to no longer refer to the SLPF AP, and the domain_name Target with corresponding Commands have been added.